requirements.txt
ecdsa
pysha3
etherscan-python
